requirements.txt
beautifulsoup4==4.4.1
chardet==2.3.0
click==6.7
html5lib==0.999
lxml==3.5.0
MarkupSafe==0.23
nose==1.3.7
olefile==0.44
Pillow==4.0.0
ranger==1.7.1
requests==2.13.0
six==1.10.0
unity-lens-photos==1.0
